# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp/NFS-e API",
        "description": "API de integração entre agendamentos e providers WhatsApp/NFS-e.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor atual"
        }
    ],
    "paths": {
...
        "/api/v1/invoices/{invoice}": {
            "get": {
                "tags": [
                    "Fiscal"
                ],
                "summary": "Consulta o status de uma NFS-e",
                "operationId": "6c4a8bbba33211c975c28f8266762ddc",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da solicitação fiscal"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
        "/api/v1/invoices/{invoice}/manual": {
            "post": {
                "tags": [
                    "Fiscal"
                ],
                "summary": "Registra uma NFS-e emitida fora do sistema (provedor manual)",
                "operationId": "761fd9ad9e36b5c33b5fe2c2637deaa4",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_number"
                                ],
                                "properties": {
                                    "invoice_number": {
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "series": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "access_key": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "protocol": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "document_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "xml": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota registrada"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Estado inválido para registro manual"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
        "/api/v1/invoices/{invoice}/cancel": {
            "post": {
                "tags": [
                    "Fiscal"
                ],
                "summary": "Cancela uma NFS-e (mantendo histórico)",
                "operationId": "545bf7234ee1ed20f63a84136c7452cf",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Erro de emissão"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota cancelada"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "422": {
                        "description": "Estado inválido para cancelamento"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "hmac": {
                "type": "apiKey",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}